Json_data/requirement.txt
appdirs==1.4.3
asgiref==3.2.7
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
distlib==0.3.0
Django==3.0
django-environ==0.4.5
django-mysql==3.5.0
djangorestframework==3.11.0
elasticsearch==7.7.0
filelock==3.0.12
mysql==0.0.2
mysql-connector-python==8.0.20
mysqlclient==1.4.6
numpy==1.18.4
pandas==1.0.3
paramiko==2.7.1
Pillow==7.1.2
pycparser==2.20
pymongo==3.10.1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
sshtunnel==0.1.5
urllib3==1.25.9
virtualenv==20.0.18
xlrd==1.2.0
